+++ src/Rocketeer/Traits/StepsRunner.php
@@ -47,10 +47,10 @@ trait StepsRunner
         $steps = $this->steps()->pullSteps();
         foreach ($steps as $step) {
             list($method, $arguments) = $step;
-            $arguments                = (array) $arguments;
+            $arguments = (array) $arguments;
 
             $results = call_user_func_array([$this, $method], $arguments);
-            $results = $results ?: $this->status();
+            $results = is_bool($results) ? $results : $this->status();
             if (!$results) {
                 return false;
             }